requirements.txt
openai
pandas
python-dotenv
langchain
streamlit
langchain_community